# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Your Project API",
    "description": "API documentation for your Node.js project.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://nodejs-hw-mongodb-wjqd.onrender.com",
      "description": "Production server on Render.com"
    }
  ],
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a new user",
        "operationId": "registerUser",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/SuccessfullyRegisteredResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "409": {
            "$ref": "#/components/responses/ConflictResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    },
...
    "/auth/send-reset-email": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send password reset email",
        "operationId": "sendResetEmail",
        "description": "Sends a password reset email to the provided email address.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendResetEmailRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResetEmailSentResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error",
                  "example": {
                    "status": 404,
                    "message": "Not Found",
                    "data": "User not found!"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to send email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error",
                  "example": {
                    "status": 500,
                    "message": "Internal Server Error",
                    "data": "Failed to send the email, please try again later."
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-pwd": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Reset user password",
        "operationId": "resetPassword",
        "description": "Reset a user's password using a valid reset token.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/PasswordResetSuccessResponse"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestResponse"
          },
          "401": {
            "description": "Token is expired or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error",
                  "example": {
                    "status": 401,
                    "message": "Unauthorized",
                    "data": "Token is expired or invalid."
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error",
                  "example": {
                    "status": 404,
                    "message": "Not Found",
                    "data": "User not found!"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerErrorResponse"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Auto-generated unique ID of the user.",
            "example": "60d5ec49f8a3c9001c9b4d8b"
          },
          "name": {
            "type": "string",
            "example": "Alice Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice.smith@example.com"
          },
          "photo": {
            "type": "string",
            "format": "uri",
            "description": "URL to the user's avatar.",
            "example": "https://example.com/avatars/alice.jpg"
          },
          "refreshToken": {
            "type": "string",
            "description": "Current refresh token for the user.",
            "nullable": true,
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-26T10:00:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-26T10:00:00Z"
          }
        }
      },
      "UserRegister": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Alice Smith"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alice.smith@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "strongpassword123"
          }
        }
      },
...
      "SendResetEmailRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "password",
          "token"
        ],
        "properties": {
          "password": {
            "type": "string",
            "format": "password",
            "example": "newStrongPassword123"
          },
          "token": {
            "type": "string",
            "description": "Password reset token received via email.",
            "example": "somerandomtokenstring"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Bad Request"
          },
          "data": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      }
    },
    "responses": {
...
      "BadRequestResponse": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error",
              "example": {
                "status": 400,
                "message": "Bad Request",
                "data": "Validation error"
              }
            }
          }
        }
      },
      "InternalServerErrorResponse": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error",
              "example": {
                "status": 500,
                "message": "Internal Server Error",
                "data": null
              }
            }
          }
        }
      },
      "SuccessfullyRegisteredResponse": {
        "description": "User registered successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "User created successfully!"
                },
                "data": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      },
      "ConflictResponse": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error",
              "example": {
                "status": 409,
                "message": "Conflict",
                "data": "Email is already in use"
              }
            }
          }
        }
      },
...
      "ResetEmailSentResponse": {
        "description": "Password reset email sent",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Password reset email sent successfully. Please check your inbox."
                },
                "data": {
                  "type": "object",
                  "nullable": true,
                  "example": null
                }
              }
            }
          }
        }
      },
      "PasswordResetSuccessResponse": {
        "description": "Password reset successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "number",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Password successfully reset!"
                },
                "data": {
                  "type": "object",
                  "nullable": true,
                  "example": null
                }
              }
            }
          }
        }
      }
    }
  }
}